Create Verbannt_PLD
annotations in lua describe in game macro's needed to run this lua, if F keys are unavailable such as on laptops. Read for further instructions. 

Massively revamped PLD lua, This lua has hooks for emntiy in spells(blue mage sub) and ja and they are used when in defense sets.
Also can correctly use FC sets for bluemagic that is listed on line 460
This set when either magical or physical defense sets are set to HP and hp is above the value set on line 466 (should be what your hp is in FC gear) to over right both precast AND midcast for cure when your ap is above this value.
All other spells will drop you into your standard fc and midcast sets for full hate / effect.
Auto movement gear is equiped when movement is detected for over about 1.5 seconds 

Note that lua on ffxiah under First and final line of defense and this one share same original lua, But this has been modified to encompass blue AoE hate Kiting Turtle and DPS.